table_test: add tests for new table option
I'm not too familiar with Go, so forgive me if I made a bunch of changes
I wasn't supposed to here.

When compiling and running it from the command line, the option would
correctly default to the value of "on" in my manual testing. However, I
couldn't for the life of me get it to behave like that in the tests
without specifically going into each one's options and explicltly adding
`WithPadColumns(PadColumnBehaviorOn)`.

I assumed it would default without that, but evidently not, and this
was the only way that the tests would work as expected.